refactor(be): 공통 `TagColor`를 `common`으로 이동 및 `ChatRoomEntity` color 컬럼 타입 수정
- `IssueTagEntity`와 `ChatRoomEntity`에서 같은 `TagColor`가 사용되어 common.enums로 이동
- `color`컬럼의 타입을 String에서 TagColor(enum)로 변경
