--- backend/src/main/java/com/ourhour/domain/project/enums/IssueTagColor.java
@@ -1,9 +1,0 @@
-package com.ourhour.domain.project.enums;
-
-public enum IssueTagColor {
-    PINK,
-    YELLOW,
-    GREEN,
-    BLUE,
-    PURPLE;
-}
