+++ backend/src/main/java/com/ourhour/global/common/enums/TagColor.java
@@ -1,0 +1,9 @@
+package com.ourhour.global.common.enums;
+
+public enum TagColor {
+    PINK,
+    YELLOW,
+    GREEN,
+    BLUE,
+    PURPLE;
+}
