+++ backend/src/test/java/com/ourhour/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.backend;
+package com.ourhour;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
